fix:
- improved testability by making setCurrentActivation return a void promise
- moved common mocking logic for 3 diffierent tests into a mocks/chrome.ts
- remediated bug within LocalStorage where null always returned since storage
object expected to be returned instead of actual value
